requirements.txt
google-auth==2.17.1
google-auth-oauthlib==1.0.0
pandas==1.5.3
gspread==5.7.2
python-dotenv==1.0.0
oauth2client==4.1.3
python-telegram-bot==20.2